+++ NEWS.md
@@ -1,5 +1,8 @@
 # S7 (development version)
 
+* `S7_inherits()` now accepts `class = NULL` to test if an object is any
+  sort of S7 object (#347).
+
 * External methods are now registered using an attribute of the S3 methods 
   table rather than an element of that environment. This prevents a warning
   being generated during the "code/documentation mismatches" check in